+++ lib/maps/mod.ts
@@ -1,4 +1,5 @@
 export * from "./model.ts";
 export * from "./parse.ts";
+export * from "./place.ts";
 export * from "./route.ts";
 export * from "./short-link.ts";